feat: implement AuthProvider for authentication context and update components to use it
